+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+* The color gutter icon now shows the correct color for color definitions
 * Add run configuration setting to run code before the main file
 
 ### Fixed